# gpt_action_openapi.yaml
openapi: 3.1.0
info:
  title: AI Day Trader Plan Approval
  version: 2.3.0
  description: Reads fresh AI/semiconductor candidates and registers explicitly approved paper plans only.
servers:
  - url: https://evict-smell-chug.ngrok-free.dev
paths:
  /v1/gpt-actions/candidates:
    get:
      operationId: getDayTradeCandidates
      summary: Read the current premarket shortlist or regular-session confirmations.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: market
          required: true
          schema: {type: string, enum: [KR, US]}
        - in: query
          name: phase
          required: false
          schema: {type: string, enum: [auto, premarket, regular], default: auto}
        - in: query
          name: limit
          required: false
          schema: {type: integer, minimum: 1, maximum: 10, default: 5}
      responses:
        '200':
          description: Read-only candidate ranking. It never arms a plan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateScanResponse'
  /v1/gpt-actions/plans:
    post:
      operationId: registerApprovedPaperTradePlan
      summary: Register a plan only after the user explicitly approves it and supplies the OTP.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradePlan'
      responses:
        '201':
          description: Plan armed for paper trading.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanReceipt'
        '409': {description: Invalid/reused OTP or duplicate plan.}
        '422': {description: Plan rejected by deterministic validation.}
  /v1/gpt-actions/plans/{plan_id}/status:
    get:
      operationId: getPaperTradePlanStatus
      summary: Check whether a submitted paper-trading plan is armed or completed.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: plan_id
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Current plan status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanStatusResponse'
  /v1/gpt-actions/experiments:
    post:
      operationId: registerPortfolioComparisonExperiment
      summary: Start three isolated paper-only cohorts after explicit user selection and OTP approval.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortfolioExperimentRequest'
      responses:
        '201':
          description: Comparison experiment started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioExperimentReceipt'
        '409': {description: Invalid/reused OTP or duplicate experiment.}
        '422': {description: Experiment rejected by deterministic validation.}
  /v1/gpt-actions/experiments/{experiment_id}:
    get:
      operationId: getPortfolioComparisonExperiment
      summary: Read allocation, portfolio, and performance for all three paper cohorts.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: experiment_id
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Current experiment comparison.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                required: [experiment_id, status, market, trade_date, candidate_symbols,
                  user_selected_symbols, cohorts, comparison]
                properties:
                  experiment_id: {type: string}
                  status: {type: string}
                  market: {type: string, enum: [KR, US]}
                  trade_date: {type: string, format: date}
                  candidate_symbols: {type: array, items: {type: string}}
                  user_selected_symbols: {type: array, items: {type: string}}
                  cohorts:
                    type: object
                    additionalProperties: false
                    required: [GPT_ALL_EQUAL, USER_FIXED_SLEEVE, USER_REALLOCATED]
                    properties:
                      GPT_ALL_EQUAL: {$ref: '#/components/schemas/CohortResult'}
                      USER_FIXED_SLEEVE: {$ref: '#/components/schemas/CohortResult'}
                      USER_REALLOCATED: {$ref: '#/components/schemas/CohortResult'}
                  comparison: {$ref: '#/components/schemas/ExperimentComparison'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    TradePlan:
      type: object
      additionalProperties: false
      required: [plan_id, plan_version, created_at, market, trade_date, expires_at,
        approval_nonce, approved_symbols]
      properties:
        plan_id: {type: string, pattern: '^[A-Za-z0-9_-]{8,80}$'}
        plan_version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        market: {type: string, enum: [KR, US]}
        trade_date: {type: string, format: date}
        expires_at: {type: string, format: date-time}
        approval_nonce: {type: string, pattern: '^\d{6}$'}
        approved_symbols:
          type: array
          minItems: 1
          maxItems: 3
          items: {$ref: '#/components/schemas/Candidate'}
    PortfolioExperimentRequest:
      type: object
      additionalProperties: false
      required: [experiment_id, created_at, market, trade_date, expires_at,
        approval_nonce, candidates, user_selected_symbols]
      properties:
        experiment_id: {type: string, pattern: '^[A-Za-z0-9_-]{8,48}$'}
        created_at: {type: string, format: date-time}
        market: {type: string, enum: [KR, US]}
        trade_date: {type: string, format: date}
        expires_at: {type: string, format: date-time}
        approval_nonce: {type: string, pattern: '^\d{6}$'}
        candidates:
          type: array
          minItems: 1
          maxItems: 3
          items: {$ref: '#/components/schemas/Candidate'}
        user_selected_symbols:
          type: array
          minItems: 1
          maxItems: 3
          items: {type: string}
    PortfolioExperimentReceipt:
      type: object
      additionalProperties: false
      required: [experiment_id, status, cohorts, message]
      properties:
        experiment_id: {type: string}
        status: {type: string, enum: [ACTIVE]}
        cohorts:
          type: array
          items:
            type: string
            enum: [GPT_ALL_EQUAL, USER_FIXED_SLEEVE, USER_REALLOCATED]
        message: {type: string}
    CohortResult:
      type: object
      additionalProperties: true
      required: [plan_status, allocation, portfolio, performance]
      properties:
        plan_status: {type: string}
        allocation: {$ref: '#/components/schemas/AllocationResult'}
        portfolio: {$ref: '#/components/schemas/ExperimentPortfolio'}
        performance: {$ref: '#/components/schemas/ExperimentPerformance'}
    AllocationResult:
      type: object
      additionalProperties: true
      required: [quantities, target_weights, estimated_invested, estimated_cash]
      properties:
        quantities:
          type: object
          additionalProperties: {type: integer}
          properties:
            example_symbol: {type: integer}
        target_weights:
          type: object
          additionalProperties: {type: number}
          properties:
            example_symbol: {type: number}
        estimated_invested: {type: number}
        estimated_cash: {type: number}
    ExperimentPortfolio:
      type: object
      additionalProperties: true
      required: [market, cash, equity, realized_pnl, positions, pending_orders]
      properties:
        market: {type: string, enum: [KR, US]}
        cash: {type: number}
        equity: {type: number}
        realized_pnl: {type: number}
        positions:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              symbol: {type: string}
              remaining: {type: integer}
        pending_orders:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              symbol: {type: string}
              state: {type: string}
    ExperimentPerformance:
      type: object
      additionalProperties: true
      required: [market, closed_trades, net_pnl, realized_net_pnl, max_drawdown_pct]
      properties:
        market: {type: string, enum: [KR, US]}
        closed_trades: {type: integer}
        net_pnl: {type: number}
        realized_net_pnl: {type: number}
        profit_factor: {type: [number, 'null']}
        max_drawdown_pct: {type: number}
    ExperimentComparison:
      type: object
      additionalProperties: true
      properties:
        user_fixed_minus_gpt_all_net_pnl: {type: number}
        user_reallocated_minus_gpt_all_net_pnl: {type: number}
        concentration_effect_net_pnl: {type: number}
        selection_uplift_per_trade: {type: [number, 'null']}
        selection_uplift_return_pct: {type: [number, 'null']}
    Candidate:
      type: object
      additionalProperties: false
      required: [symbol, exchange, reason, strategy_type, entry, stop_loss, take_profit,
        force_exit_time]
      properties:
        symbol: {type: string}
        exchange: {type: string}
        reason: {type: string}
        strategy_type: {type: string, enum: [rules, pullback_rebreak]}
        entry: {$ref: '#/components/schemas/Entry'}
        stop_loss:
          type: object
          additionalProperties: false
          required: [price]
          properties:
            price: {type: number, exclusiveMinimum: 0}
            execution_type: {type: string, enum: [marketable_limit], default: marketable_limit}
            limit_offset_pct: {type: number, minimum: 0.05, maximum: 2, default: 0.2}
            emergency_exit_after_sec: {type: number, minimum: 0.1, maximum: 30, default: 2}
        take_profit:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: object
            additionalProperties: false
            required: [price, quantity_pct]
            properties:
              price: {type: number, exclusiveMinimum: 0}
              quantity_pct: {type: integer, minimum: 1, maximum: 100}
        force_exit_time: {type: string, format: time}
        exit_policy: {$ref: '#/components/schemas/ExitPolicy'}
        pullback_rebreak: {$ref: '#/components/schemas/PullbackRebreak'}
        premarket_guard: {$ref: '#/components/schemas/PremarketGuard'}
    Entry:
      type: object
      additionalProperties: false
      required: [trigger_price, limit_price, start_time, end_time, price_only, rules]
      properties:
        trigger_price: {type: number, exclusiveMinimum: 0}
        limit_price: {type: number, exclusiveMinimum: 0}
        start_time: {type: string, format: time}
        end_time: {type: string, format: time}
        price_only: {type: boolean}
        rules: {$ref: '#/components/schemas/RuleGroup'}
    RuleGroup:
      type: object
      additionalProperties: false
      required: [mode, predicates, groups]
      properties:
        mode: {type: string, enum: [all, any]}
        predicates:
          type: array
          maxItems: 12
          items: {$ref: '#/components/schemas/Predicate'}
        groups:
          type: array
          maxItems: 4
          items: {$ref: '#/components/schemas/RuleGroupLevel2'}
    RuleGroupLevel2:
      type: object
      additionalProperties: false
      required: [mode, predicates, groups]
      properties:
        mode: {type: string, enum: [all, any]}
        predicates:
          type: array
          maxItems: 12
          items: {$ref: '#/components/schemas/Predicate'}
        groups:
          type: array
          maxItems: 4
          items: {$ref: '#/components/schemas/RuleGroupLevel3'}
    RuleGroupLevel3:
      type: object
      additionalProperties: false
      required: [mode, predicates, groups]
      properties:
        mode: {type: string, enum: [all, any]}
        predicates:
          type: array
          maxItems: 12
          items: {$ref: '#/components/schemas/Predicate'}
        groups:
          type: array
          maxItems: 4
          items: {$ref: '#/components/schemas/RuleGroupLevel4'}
    RuleGroupLevel4:
      type: object
      additionalProperties: false
      required: [mode, predicates, groups]
      properties:
        mode: {type: string, enum: [all, any]}
        predicates:
          type: array
          maxItems: 12
          items: {$ref: '#/components/schemas/Predicate'}
        groups:
          type: array
          maxItems: 0
          items: {type: string}
    Predicate:
      type: object
      additionalProperties: false
      required: [indicator, operator, value]
      properties:
        indicator:
          type: string
          enum: [last, bid, ask, spread_pct, regular_open_price, premarket_open,
            premarket_high, premarket_low, premarket_last, premarket_vwap,
            premarket_volume, premarket_gap_pct, previous_open, previous_high,
            previous_low, previous_close, opening_range_5_high, opening_range_5_low,
            opening_range_10_high, opening_range_10_low, opening_range_15_high,
            opening_range_15_low, vwap_regular, atr_14_1m_regular,
            ema_9_1m_regular, ema_20_1m_regular, ema_50_1m_regular,
            rsi_14_1m_regular, relative_volume_cumulative_20d_same_time_regular,
            gap_pct, market_above_vwap_regular, recent_high_5_1m_regular,
            recent_low_5_1m_regular, bar_volume_1m_regular]
        operator: {type: string, enum: [eq, ne, gt, gte, lt, lte, cross_above, cross_below]}
        value: {oneOf: [{type: number}, {type: boolean}, {type: string}]}
        confirm_ticks: {type: integer, minimum: 1, maximum: 20, default: 1}
        hold_above_ms: {type: integer, minimum: 0, maximum: 60000, default: 0}
        minimum_cross_pct: {type: number, minimum: 0, maximum: 5, default: 0}
        cooldown_sec: {type: integer, minimum: 0, maximum: 3600, default: 0}
    ExitPolicy:
      type: object
      additionalProperties: false
      properties:
        max_holding_minutes: {type: integer, minimum: 1, maximum: 360, default: 90}
        no_progress_exit_minutes: {type: integer, minimum: 1, maximum: 180, default: 30}
        no_progress_min_pct: {type: number, minimum: 0, maximum: 5, default: 0.3}
        exit_if_below_vwap_sec: {type: integer, minimum: 1, maximum: 600, default: 60}
        move_stop_to_entry_after_tp1: {type: boolean, default: false}
    PullbackRebreak:
      type: object
      additionalProperties: false
      properties:
        breakout_confirm_ticks: {type: integer, minimum: 1, maximum: 20, default: 3}
        breakout_hold_ms: {type: integer, minimum: 0, maximum: 60000, default: 1000}
        breakout_relative_volume_min: {type: number, minimum: 0.1, maximum: 10, default: 1.5}
        pullback_depth_atr_min: {type: number, minimum: 0.05, maximum: 3, default: 0.2}
        pullback_depth_atr_max: {type: number, minimum: 0.1, maximum: 5, default: 0.8}
        pullback_min_minutes: {type: number, minimum: 0.5, maximum: 30, default: 2}
        pullback_max_minutes: {type: number, minimum: 1, maximum: 60, default: 10}
        pullback_rsi_min: {type: number, minimum: 0, maximum: 100, default: 45}
        pullback_rsi_max: {type: number, minimum: 0, maximum: 100, default: 60}
        pullback_volume_ratio_max: {type: number, exclusiveMinimum: 0, maximum: 2, default: 0.7}
        support_breach_pct_max: {type: number, minimum: 0, maximum: 2, default: 0.15}
        rebreak_rsi_min: {type: number, minimum: 0, maximum: 100, default: 50}
        rebreak_rsi_max: {type: number, minimum: 0, maximum: 100, default: 68}
        rebreak_relative_volume_min: {type: number, minimum: 0.1, maximum: 10, default: 1.3}
        spread_pct_max: {type: number, exclusiveMinimum: 0, maximum: 2, default: 0.15}
        require_market_above_vwap: {type: boolean, default: true}
    PremarketGuard:
      type: object
      additionalProperties: false
      required: [reference_price]
      properties:
        reference_price: {type: number, exclusiveMinimum: 0}
        max_open_deviation_pct: {type: number, exclusiveMinimum: 0, maximum: 10, default: 1}
        max_spread_pct: {type: number, exclusiveMinimum: 0, maximum: 2, default: 0.15}
        relative_volume_min: {type: number, exclusiveMinimum: 0, maximum: 10, default: 1.3}
        require_above_vwap: {type: boolean, default: true}
        require_market_above_vwap: {type: boolean, default: true}
    CandidateScanResponse:
      type: object
      additionalProperties: false
      required: [market, trade_date, phase, status, as_of, window, execution_note,
        candidates, exclusions]
      properties:
        market: {type: string, enum: [KR, US]}
        trade_date: {type: string, format: date}
        phase: {type: string, enum: [premarket, regular, closed]}
        status: {type: string}
        as_of: {type: string, format: date-time}
        window:
          type: object
          additionalProperties: true
          properties:
            status: {type: string}
            phase: {type: string}
            premarket_scan_start: {type: string, format: date-time}
            candidate_selection_end: {type: string, format: date-time}
            regular_open: {type: string, format: date-time}
            regular_confirmation_at: {type: string, format: date-time}
            regular_close: {type: string, format: date-time}
        execution_note: {type: string}
        candidates:
          type: array
          items:
            type: object
            additionalProperties: true
            required: [symbol, name, exchange, score, as_of, session, quote_scope,
              last, bid, ask, spread_pct, indicators, premarket_guard_template]
            properties:
              symbol: {type: string}
              name: {type: string}
              exchange: {type: string}
              score: {type: number}
              as_of: {type: string, format: date-time}
              session: {type: string}
              quote_scope: {type: string}
              last: {type: number}
              bid: {type: number}
              ask: {type: number}
              spread_pct: {type: number}
              bid_size: {type: integer}
              ask_size: {type: integer}
              indicators:
                type: object
                additionalProperties: true
                properties: {}
              premarket_guard_template: {$ref: '#/components/schemas/PremarketGuard'}
        exclusions:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [symbol, reasons]
            properties:
              symbol: {type: string}
              reasons: {type: array, items: {type: string}}
    PlanReceipt:
      type: object
      additionalProperties: false
      required: [plan_id, status, content_hash, message]
      properties:
        plan_id: {type: string}
        status: {type: string}
        content_hash: {type: string}
        message: {type: string}
    PlanStatusResponse:
      type: object
      required: [plan_id, plan_version, market, trade_date, status, content_hash]
      properties:
        plan_id: {type: string}
        plan_version: {type: integer}
        market: {type: string, enum: [KR, US]}
        trade_date: {type: string, format: date}
        status:
          type: string
          enum: [VALIDATED, ARMED, RUNNING, COMPLETED, EXPIRED, REJECTED, CANCELLED, RISK_BLOCKED]
        content_hash: {type: string}